chore(Flow): extract state section hooks [#1752]
