+++ packages/ui/src/ui/pages/flow/flow-components/FlowStateViewer/FlowStateSection/FlowStateSection.tsx
@@ -2,44 +2,25 @@ import React from 'react';
 import cn from 'bem-cn-lite';
 
 import {Alert} from '@gravity-ui/uikit';
-import {skipToken} from '@reduxjs/toolkit/query';
 
 import {type tanstack} from '../../../../../components/DataTableGravity';
 import WithStickyToolbar from '../../../../../components/WithStickyToolbar/WithStickyToolbar';
-import {useDebouncedValue} from '../../../../../hooks/useDebouncedValue';
 import {YTApiId} from '../../../../../rum/rum-wrap-api';
 import {useCheckPermissionQuery} from '../../../../../store/api/yt/checkPermissions';
-import {useFlowReadStatesQuery, useFlowStaticSpecQuery} from '../../../../../store/api/yt/flow';
-import {FlowTab} from '../../../../../store/reducers/flow/filters';
-import {selectCluster, selectCurrentUserName} from '../../../../../store/selectors/global';
-import {makeFlowLink} from '../../../../../utils/app-url/makeFlowLink';
+import {selectCurrentUserName} from '../../../../../store/selectors/global';
 
 import {useSelector} from '../../../../../store/redux-hooks';
 
 import {FlowDeleteStatesDialog} from '../FlowDeleteStatesDialog/FlowDeleteStatesDialog';
 import {FlowStateFilters} from '../FlowStateFilters/FlowStateFilters';
 import {FlowStateResults} from '../FlowStateResults/FlowStateResults';
-import {
-    AUTO_LOAD_DEBOUNCE_MS,
-    buildStateReadBody,
-    flattenReadStatesResponse,
-    selectDeletableRows,
-} from '../state-requests';
-import {
-    getComputationStateNames,
-    resolveKeySchema,
-    resolveRowKeySchema,
-    seedStateFilters,
-} from '../state-filters';
-import {buildRowFilterUpdate, resolveStateStoragePath} from '../state-values';
+import {flattenReadStatesResponse, selectDeletableRows} from '../state-requests';
+import {seedStateFilters} from '../state-filters';
 import {isWriteDeniedByPermission} from '../state-delete';
+import {useFlowStateCellHandlers} from './use-flow-state-cell-handlers';
+import {useFlowStateRead} from './use-flow-state-read';
 import i18n from './i18n';
-import type {
-    FlowStateCellHandlers,
-    FlowStateFiltersValue,
-    FlowStateResultRow,
-    FlowStateRowFilterField,
-} from '../types';
+import type {FlowStateFiltersValue, FlowStateResultRow} from '../types';
 
 import './FlowStateSection.scss';
 
@@ -56,51 +37,28 @@ export function FlowStateSection({
     fixedComputationId,
     initialFilters,
 }: FlowStateSectionProps) {
-    const [filters, setFilters] = React.useState<FlowStateFiltersValue>(() =>
-        seedStateFilters(fixedComputationId, initialFilters),
-    );
     const [rowSelection, setRowSelection] = React.useState<tanstack.RowSelectionState>({});
     const [deleteRows, setDeleteRows] = React.useState<Array<FlowStateResultRow>>();
     const [deleteVisible, setDeleteVisible] = React.useState(false);
     const login = useSelector(selectCurrentUserName);
-    const cluster = useSelector(selectCluster);
 
-    const {data: staticSpec} = useFlowStaticSpecQuery({parameters: {pipeline_path}});
+    const {
+        filters,
+        setFilters,
+        staticSpec,
+        hasScope,
+        validationError,
+        response,
+        loading,
+        error,
+        refetch,
+    } = useFlowStateRead({pipeline_path, fixedComputationId, initialFilters});
+
     const permissionResult = useCheckPermissionQuery({
         id: YTApiId.checkPermissions,
         parameters: {path: pipeline_path, user: login, permission: 'write'},
     });
     const writeDenied = isWriteDeniedByPermission(permissionResult);
-
-    const hasScope = Boolean(filters.computationId || filters.partitionId);
-    const debouncedFilters = useDebouncedValue(filters, AUTO_LOAD_DEBOUNCE_MS);
-    const hasDebouncedScope = Boolean(
-        debouncedFilters.computationId || debouncedFilters.partitionId,
-    );
-    const built = hasDebouncedScope
-        ? buildStateReadBody(
-              debouncedFilters,
-              resolveKeySchema(
-                  staticSpec,
-                  debouncedFilters.computationId,
-                  debouncedFilters.stateName,
-                  debouncedFilters.target,
-              ),
-          )
-        : undefined;
-    const validationError =
-        built && 'error' in built ? i18n(built.error.errorKey, built.error.params) : undefined;
-    const {
-        data,
-        error,
-        isFetching: loading,
-        refetch,
-    } = useFlowReadStatesQuery(
-        built && 'body' in built
-            ? {parameters: {pipeline_path}, body: {...built.body, limit: debouncedFilters.limit}}
-            : skipToken,
-    );
-    const response = hasDebouncedScope ? data : undefined;
 
     React.useEffect(() => {
         setRowSelection({});
@@ -112,54 +70,25 @@ export function FlowStateSection({
         [rows, rowSelection],
     );
 
-    const handleFiltersChange = React.useCallback((next: FlowStateFiltersValue) => {
-        setFilters(next);
-        setRowSelection({});
-    }, []);
+    const handleFiltersChange = React.useCallback(
+        (next: FlowStateFiltersValue) => {
+            setFilters(next);
+            setRowSelection({});
+        },
+        [setFilters],
+    );
 
     const handleReset = () => {
         handleFiltersChange(seedStateFilters(fixedComputationId, initialFilters));
     };
 
-    const getRowFilterUpdate = React.useCallback(
-        (row: FlowStateResultRow, field: FlowStateRowFilterField) =>
-            buildRowFilterUpdate(filters, row, field, {
-                ...resolveRowKeySchema(staticSpec, row),
-                stateNames: getComputationStateNames(
-                    staticSpec,
-                    filters.computationId,
-                    row.section === 'joined_external_key_state' ? 'all' : row.section,
-                ),
-                fixedComputationId,
-            }),
-        [filters, staticSpec, fixedComputationId],
-    );
-
-    const resolveStoragePath = React.useCallback(
-        (row: FlowStateResultRow) => resolveStateStoragePath(row, pipeline_path, staticSpec),
-        [pipeline_path, staticSpec],
-    );
-
-    const resolveComputationLink = React.useCallback(
-        (computationId: string) =>
-            makeFlowLink({
-                path: pipeline_path,
-                cluster,
-                tab: FlowTab.COMPUTATIONS,
-                computation: computationId,
-            }),
-        [pipeline_path, cluster],
-    );
-
-    const cellHandlers = React.useMemo<FlowStateCellHandlers>(
-        () => ({
-            getRowFilterUpdate,
-            onFiltersChange: handleFiltersChange,
-            resolveStoragePath,
-            resolveComputationLink,
-        }),
-        [getRowFilterUpdate, handleFiltersChange, resolveStoragePath, resolveComputationLink],
-    );
+    const cellHandlers = useFlowStateCellHandlers({
+        pipeline_path,
+        filters,
+        staticSpec,
+        fixedComputationId,
+        onFiltersChange: handleFiltersChange,
+    });
 
     const handleDeleteSelected = () => {
         setDeleteRows(selectedRows);
